+++ packages/plugin-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 18.0.0-rc.8
+
+### Patch Changes
+
+- 0016d7ff: Added node, serverClient and config presets
+
 ## 18.0.0-rc.7
 
 ### Major Changes
@@ -7,7 +13,7 @@
 - cc9894d2: Plugin config now has completely different configuration definition
   Link, dev and build commands are now more efficient and use less watchers for multiple bundle options
   Transformers option is now deprecated (they are part of the core replaced with type and bundle option)
-  New CLI argument --clientServerBundle
+  New CLI argument --clientServerConfig
 
 ## 18.0.0-rc.6